[fix] 화면 자동 랜더링 수정
